Add docs, chatbot and prompt/response formats
Update the readme.md describing how to use the library and its keywords.
Add chatbot which can respond to your messages with an optional flag to keep history.
Create a format for both prompts and responses which act as a standard for modules communicating with
AI models and vice versa.