Add examples for keyword usage + some minor fixes
Add examples on how to use the Generate Test Data and Generate Response keywords aswell as setters.
Uncommented a line that installs the necessary libraries to use this library.
Add types to the keywords arguments so numbers aren't seen as strings.
Update error message in case AI response isn't json-formatable and small changes to system prompt to control response format.